pyproject.toml
[project]
name = "hello-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi[standard]>=0.135.2",
    "greenlet>=3.3.2",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "sqlmodel>=0.0.38",
]
